PEN-802 small manual promo refactoring
