--- blocks/small-manual-promo-block/features/small-manual-promo/_children/promo_link.jsx
@@ -1,18 +1,0 @@
-import React, { useCallback } from 'react';
-
-const usePromoElementLink = (customFields) => useCallback((element, props) => (
-  <a
-    href={customFields.linkURL || '#'}
-    className={(props && props.className) || ''}
-    title={customFields.headline}
-    target={customFields.newTab ? '_blank' : '_self'}
-    rel={customFields.newTab ? 'noreferrer noopener' : ''}
-    onClick={!customFields.linkURL ? (evt) => {
-      evt.preventDefault();
-    } : undefined}
-  >
-    {element}
-  </a>
-), [customFields.linkURL, customFields.headline, customFields.newTab]);
-
-export default usePromoElementLink;
